refactor(desktop): split agents page shell orchestration
